+++ src/style/dump/dump_grid_gz.cpp
@@ -91,7 +91,7 @@ DumpGridGz::~DumpGridGz() {
 void DumpGridGz::write() {
 
   int ithread;
-  pair<thread, vector<double>> *th = nullptr;
+  pair<thread, vector<float>> *th = nullptr;
 
   for (int i=0; i<threads.size(); i++) {
     if (threads[i].second.empty()) {
@@ -108,7 +108,7 @@ void DumpGridGz::write() {
     th = &threads.back();
   }
 
-  vector<double> &buf = th->second;
+  vector<float> &buf = th->second;
 
   // Open dump file:
   size_t pos_asterisk = filename.find('*');
@@ -258,7 +258,7 @@ void DumpGridGz::write_to_file(bigint i, string fdump, bigint total_nn, bigint t
 
   int size_one = output_var.size();
 
-  vector<double> &buf = threads[i].second;
+  vector<float> &buf = threads[i].second;
 
   int m = 0;
   for (int j = 0; j < total_nn; j++) {
