+++ src/style/dump/dump_particle_gz.cpp
@@ -128,7 +128,7 @@ DumpParticleGz::~DumpParticleGz() {
 void DumpParticleGz::write() {
 
   int ithread;
-  pair<thread, vector<double>> *th = nullptr;
+  pair<thread, vector<float>> *th = nullptr;
 
   for (int i=0; i<threads.size(); i++) {
     if (threads[i].second.empty()) {
@@ -145,7 +145,7 @@ void DumpParticleGz::write() {
     th = &threads.back();
   }
 
-  vector<double> &buf = th->second;
+  vector<float> &buf = th->second;
 
   // Open dump file:
   size_t pos_asterisk = filename.find('*');
@@ -363,7 +363,7 @@ void DumpParticleGz::write_to_file(bigint i, string fdump, bigint total_np, bigi
 
   int size_one = output_var.size();
 
-  vector<double> &buf = threads[i].second;
+  vector<float> &buf = threads[i].second;
 
   int m = 0;
   for (int j = 0; j < total_np; j++) {
