+++ src/style/expression/core/operation/operator/expression_operators_arithmetic.h
@@ -50,7 +50,7 @@ class ExpressionNegation:
   public ExpressionOperation<ExpressionNegation, false, false, 3, 1>
 {
 public:
-  KOKKOS_INLINE_FUNCTION double
+  KOKKOS_INLINE_FUNCTION float
   evaluate(int i) const
   {
     return -get_value(0, i);
